backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
assemblyai==0.34.0
azure-core==1.31.0
azure-cosmos==4.7.0
azure-identity==1.19.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.23.1
bidict==0.23.1
blinker==1.8.2
bump2version==1.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
msal==1.31.0
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
